+++ cobaya/__init__.py
@@ -1,5 +1,5 @@
 __author__ = "Jesus Torrado and Antony Lewis"
-__version__ = "1.0.4"
-__obsolete__ = True
+__version__ = "1.1beta"
+__obsolete__ = False
 __year__ = "2019"
 __url__ = "https://cobaya.readthedocs.io"