+++ cobaya/grid_tools/runbatch.py
@@ -5,9 +5,10 @@ import hashlib
 import os
 
 from cobaya.grid_tools import batchjob_args, jobqueue
-
+from cobaya.tools import warn_deprecation
 
 def run():
+    warn_deprecation()
     Opts = batchjob_args.batchArgs('Submit jobs to run chains or importance sample',
                                    notExist=True, notall=True, converge=True)
     jobqueue.addArguments(Opts.parser, combinedJobs=True)