+++ cobaya/likelihoods/_planck_clik_prototype/_planck_clik_prototype.py
@@ -56,7 +56,6 @@ of their length.
 Installation
 ------------
 
-
 This likelihood can be installed automatically as explained in :doc:`installation_cosmo`.
 If are following the instructions there (you should!), you don't need to read the rest
 of this section.
@@ -71,8 +70,8 @@ of this section.
    ``sudo apt install liblapack3 liblapack-dev``).
 
    If you want to re-compile the Planck likelihood to your liking (e.g. with MKL), simply
-   go into the chosen modules installation folder and re-run the ``./waf configure`` and
-   ``./waf install`` with the desired options.
+   go into the chosen modules installation folder and re-run the ``python waf configure``
+   and ``python waf install`` with the desired options.
 
 However, if you wish to install it manually or have a previous installation already in
 your system, simply take note of the path to the ``plc-2.0`` and ``plc_2.0`` folders and
@@ -86,8 +85,16 @@ mention it below each Planck likelihood as
      planck_2015_plikHM_TTTEEE:
        path: /path/to/planck_2015
 
+
 Manual installation of Planck 2015 likelihoods
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. warning::
+
+   For the time being (waiting for Planck 2018's data release), use instead the
+   alternative 'clik' code at
+   `<https://cdn.cosmologist.info/cosmobox/plc-2.1_py3.tar.bz2>`_, which is
+   compatible with Python 3 and gcc `>5`.
 
 Assuming you are installing all your likelihoods under ``/path/to/likelihoods``:
 
@@ -100,7 +107,7 @@ Assuming you are installing all your likelihoods under ``/path/to/likelihoods``:
    $ tar xvjf data-action?COSMOLOGY.COSMOLOGY_OID=1904
    $ rm data-action?COSMOLOGY.COSMOLOGY_OID=1904
    $ cd plc-2.0
-   $ ./waf configure # options
+   $ python waf configure  # [options]
 
 If the last step failed, try adding the option ``--install_all_deps``.
 It it doesn't solve it, follow the instructions in the ``readme.md``
@@ -109,8 +116,8 @@ file in the ``plc-2.0`` folder.
 If you have Intel's compiler and Math Kernel Library (MKL), you may want to also add the
 option ``--lapack_mkl=${MKLROOT}`` in the last line to make use of it.
 
-If ``./waf configure`` ended successfully run ``./waf install`` in the same folder.
-You do **not** need to run ``clik_profile.sh``, as advised.
+If ``python waf configure`` ended successfully run ``python waf install``
+in the same folder. You do **not** need to run ``clik_profile.sh``, as advised.
 
 Now, download the required likelihood files from the
 `Planck Legacy Archive <https://pla.esac.esa.int/pla/#cosmology>`_ (Europe) or the
@@ -118,10 +125,10 @@ Now, download the required likelihood files from the
 
 For instance, if you want to reproduce the baseline Planck 2015 results,
 download the file ``COM_Likelihood_Data-baseline_R2.00.tar.gz``
-from any of the two links above, and uncompress it under the ``planck_2015`` folder
+from any of the two links above, and decompress it under the ``planck_2015`` folder
 that you created above.
 
-Finally, download and uncompress in the ``planck_2015`` folder the last release at
+Finally, download and decompress in the ``planck_2015`` folder the last release at
 `this repo <https://github.com/CobayaSampler/planck_supp_data_and_covmats/releases>`_.
 
 """
@@ -143,10 +150,7 @@ from cobaya.likelihood import Likelihood
 from cobaya.log import HandledException
 from cobaya.conventions import _path_install, _likelihood
 from cobaya.input import get_default_info
-from cobaya.install import pip_install
-
-# To see full clik build output even if installs OK (e.g. to check warnings)
-_clik_verbose = 'clik' in os.getenv('TRAVIS_COMMIT_MESSAGE', '')
+from cobaya.install import pip_install, download_file
 
 
 class _planck_clik_prototype(Likelihood):
@@ -176,7 +180,6 @@ class _planck_clik_prototype(Likelihood):
             # test and import clik
             is_installed_clik(self.path_clik, log_and_fail=True, import_it=False)
             import clik
-
         # Loading the likelihood data
         if not os.path.isabs(self.clik_file):
             self.path_data = getattr(self, "path_data", os.path.join(
@@ -253,61 +256,11 @@ class _planck_clik_prototype(Likelihood):
 # path to be shared by all Planck likelihoods
 common_path = "planck_2015"
 
-
-def download_file(file, path, no_progress_bars=False, name=None):
-    log = logging.getLogger(name or __name__)
-    try:
-        from wget import download, bar_thermometer
-        wget_kwargs = {"out": path, "bar":
-            (bar_thermometer if not no_progress_bars else None)}
-        filename = download(file, **wget_kwargs)
-        print('Got: %s' % filename)
-    except:
-        log.error("Error downloading!")
-        return False
-    finally:
-        print("")
-    # uncompress
-    import os
-    import tarfile
-    extension = os.path.splitext(filename)[-1][1:]
-    tar = tarfile.open(filename, "r:" + extension)
-    try:
-        tar.extractall(path)
-        tar.close()
-        os.remove(filename)
-        return True
-    except:
-        log.error("Error decompressing downloaded file! Corrupt file?)")
-        return False
-
-
-def download_from_planck(product_id, path, no_progress_bars=False, name=None):
-    log = logging.getLogger(name or __name__)
-    try:
-        from wget import download, bar_thermometer
-        wget_kwargs = {"out": path, "bar":
-            (bar_thermometer if not no_progress_bars else None)}
-        prefix = r"https://pla.esac.esa.int/pla-sl/data-action?COSMOLOGY.COSMOLOGY_OID="
-        filename = download(prefix + product_id, **wget_kwargs)
-    except:
-        log.error("Error downloading!")
-        return False
-    finally:
-        print("")
-    # uncompress
-    import os
-    import tarfile
-    extension = os.path.splitext(filename)[-1][1:]
-    tar = tarfile.open(filename, "r:" + extension)
-    try:
-        tar.extractall(path)
-        tar.close()
-        os.remove(filename)
-        return True
-    except:
-        log.error("Error decompressing downloaded file! Corrupt file?)")
-        return False
+# To see full clik build output even if installs OK (e.g. to check warnings)
+_clik_verbose = any(
+    [(s in os.getenv('TRAVIS_COMMIT_MESSAGE', '')) for s in ["clik", "planck"]])
+# Don't try again to install clik if it failed for a previous likelihood
+_clik_install_failed = False
 
 
 def is_installed_clik(path, log_and_fail=False, import_it=True):
@@ -363,25 +316,22 @@ def execute(command):
 
 def install_clik(path, no_progress_bars=False):
     log = logging.getLogger("clik")
+    log.info("Installing pre-requisites...")
     for req in ("cython", "astropy"):
-        from importlib import import_module
-        try:
-            import_module(req)
-        except ImportError:
-            log.info("clik: installing requisite '%s'...", req)
-            exit_status = pip_install(req)
-            if exit_status:
-                log.error("Failed installing '%s'.", req)
-                raise HandledException
-    log.info("clik: downloading...")
-    if not download_file('https://cdn.cosmologist.info/cosmobox/plc-2.1_py3.tar.bz2', path,
-                         no_progress_bars=no_progress_bars, name="clik"):
+        exit_status = pip_install(req)
+        if exit_status:
+            log.error("Failed installing '%s'.", req)
+            raise HandledException
+    log.info("Downloading...")
+    click_url = 'https://cdn.cosmologist.info/cosmobox/plc-2.1_py3.tar.bz2'
+    if not download_file(click_url, path, decompress=True,
+                         no_progress_bars=no_progress_bars, logger=log):
         log.error("Not possible to download clik.")
         return False
     source_dir = os.path.join(path, os.listdir(path)[0])
-    print('clik: installing from directory %s' % source_dir)
+    log.info('Installing from directory %s' % source_dir)
     if True:  # should be fixed
-        log.info("clik: patching origin of cfitsio")
+        log.info("Patching origin of cfitsio")
         cfitsio_filename = os.path.join(source_dir, "waf_tools", "cfitsio.py")
         with open(cfitsio_filename, "r") as cfitsio_file:
             lines = cfitsio_file.readlines()
@@ -392,22 +342,21 @@ def install_clik(path, no_progress_bars=False):
                 "'cfitsio3280.tar.gz')\n")
         with open(cfitsio_filename, "w") as cfitsio_file:
             cfitsio_file.write("".join(lines))
-
     cwd = os.getcwd()
     try:
         os.chdir(source_dir)
-        log.info("clik: configuring... (and maybe installing dependencies...)")
-        if not execute(["./waf", "configure", "--install_all_deps"]):
+        log.info("Configuring... (and maybe installing dependencies...)")
+        if not execute([sys.executable, "waf", "configure", "--install_all_deps"]):
             log.error("Configuration failed!")
             return False
 
-        log.info("clik: compiling...")
-        if not execute(["./waf", "install"]):
+        log.info("Compiling...")
+        if not execute([sys.executable, "waf", "install"]):
             log.error("Compilation failed!")
             return False
     finally:
         os.chdir(cwd)
-    log.info("clik: finished!")
+    log.info("Finished!")
     return True
 
 
@@ -430,9 +379,6 @@ def is_installed(**kwargs):
             is_installed as is_installed_supp
         result &= is_installed_supp(**kwargs)
     return result
-
-
-_clik_install_failed = False
 
 
 def install(path=None, name=None, force=False, code=True, data=True,
@@ -467,10 +413,12 @@ def install(path=None, name=None, force=False, code=True, data=True,
         if force or not is_installed(path=path, name=name, code=False, data=True):
             # Extract product_id
             product_id, _ = get_product_id_and_clik_file(name)
-            # Download and uncompress the particular likelihood
+            # Download and decompress the particular likelihood
             log.info("Downloading likelihood data...")
-            if not download_from_planck(product_id, paths["data"],
-                                        no_progress_bars=no_progress_bars, name=name):
+            prefix = (r"https://pla.esac.esa.int/pla-sl/"
+                      "data-action?COSMOLOGY.COSMOLOGY_OID=")
+            if not download_file(prefix + product_id, paths["data"], decompress=True,
+                                 logger=log, no_progress_bars=no_progress_bars):
                 log.error("Not possible to download this likelihood.")
                 success = False
             # Additional data and covmats